+++ lib/src/tenvio/src/package.dart
@@ -6,13 +6,6 @@ class TenvioPackage with _$TenvioPackage {
     /// [id] is the unique identifier of the package
     required String id,
 
-    /// [packerId] is the unique identifier of the operator that will pack the package.
-    String? packerId,
-
-    /// [packer] is the operator that will pack the package.
-    /// This value will only be filled when the package is assigned to an operator.
-    User? packer,
-
     /// [trackingId] is the unique identifier of the tracking of the package.
     /// The equivalent to the tracking number of a package.
     required String trackingId,
@@ -26,6 +19,24 @@ class TenvioPackage with _$TenvioPackage {
     /// [items] is the list of items that are part of the package.
     List<TenvioPackageQuantity>? items,
 
+    /// [destinationType] Type of the destination.
+    @TenvioDestinationTypeOrNullConverter() TenvioDestinationType? destinationType,
+
+    /// [destinationWarehouse] Warehouse destination of the order.
+    Asset? destinationWarehouse,
+
+    /// [destinationWarehouseId] ID of the warehouse destination of the order.
+    String? destinationWarehouseId,
+
+    /// [destinationUser] User destination of the order.
+    User? destinationUser,
+
+    /// [destinationUserId] ID of the user destination of the order.
+    String? destinationUserId,
+
+    /// [destinationUnregisteredUser] Unregistered user destination of the order.
+    TenvioUnregisteredUser? destinationUnregistered,
+
     /// [status] is the status of the package.
     @TenvioPackageStatusConverter() required TenvioPackageStatus status,
 
@@ -34,9 +45,6 @@ class TenvioPackage with _$TenvioPackage {
 
     /// [updatedAt] is the date when the package was last updated.
     @TimestampConverter() required DateTime updatedAt,
-
-    /// [totalItems] is the total number of items in the package.
-    int? totalItems,
   }) = _TenvioPackage;
 
   factory TenvioPackage.fromJson(Map<String, dynamic> json) => _$TenvioPackageFromJson(json);
@@ -45,9 +53,6 @@ class TenvioPackage with _$TenvioPackage {
 @freezed
 class TenvioPackageQuantity with _$TenvioPackageQuantity {
   const factory TenvioPackageQuantity({
-    /// [id] is the unique identifier of the package quantity
-    required String id,
-
     /// [matrixId] is the unique identifier of the matrix that the package quantity is related.
     required String matrixId,